В org.cherchgk.actions.user.DeleteUserAction.execute добавлена проверка прав на удаление пользователя.

+++ src/org/cherchgk/actions/user/DeleteUserAction.java
@@ -2,6 +2,7 @@ package org.cherchgk.actions.user;
 
 import com.opensymphony.xwork2.Action;
 import com.opensymphony.xwork2.ActionSupport;
+import org.apache.shiro.SecurityUtils;
 import org.cherchgk.services.SecurityService;
 import org.cherchgk.utils.ActionContextHelper;
 
@@ -21,7 +22,9 @@ public class DeleteUserAction extends ActionSupport {
     @Override
     public String execute() throws Exception {
         Long userId = Long.valueOf(ActionContextHelper.getRequestParameterValue("user.id"));
-        securityService.deleteUser(userId);
+        if (SecurityUtils.getSubject().isPermitted("user:edit:" + userId)) {
+            securityService.deleteUser(userId);
+        }
         return Action.SUCCESS;
     }
 }
